refactor: move Introduction component to shared directory inside features and update imports and move about inside features
