+++ src/features/shared/AdditionalInfo.tsx
